Add files via upload
collatz conjecture can be defined as an undirected graph, not directly related to prime numbers. Right?

Here is my thinking


I try to show that the collatz orbits can be specified to an undirected graph, from which any legal collatz chain can be constructed, and that we can obtain several analytic expressions that constitute segments.

Then, according to the properties of the analytical formula, it can be concluded that there is no continuous climbing orbit in the collatz orbit.

Then it is proved by induction that the orbitals of continuous climbing must be equal to R1 and R0, according to the fact that there cannot be two consecutive even nodes, and according to the drawer principle.

Now we can construct a long enough collatz orbit diagram. But it doesn't climb indefinitely,
According to the odd and even rules, it must climb in a zigzag shape, and R1 rules decrease rapidly as the orbit grows.

We analyzed the moment the orbit turns, and R0 is always increasing,

Basically, the idea is to call the odd rule R1, the even rule R0, and then expand on those combinations to logically satisfy the derivation.

I look forward to pointing out the errors and will add your suggestions